+++ src/utils/createQuizData.js
@@ -1,4 +1,4 @@
-
+import { shuffle } from "./array/shuffle"
 
 export const createQuizData = (data) => {
     console.log('a', data)
@@ -31,21 +31,4 @@ export const createQuizData = (data) => {
     })
     console.log(list)
     return list
-}
-
-const shuffle = (array) => {
-    let newArray = structuredClone(array)
-    let currentIndex = array.length
-    let temporaryValue
-    let randomIndex
-
-    while (0 !== currentIndex) {
-        randomIndex = Math.floor(Math.random() * currentIndex)
-        currentIndex -= 1
-        temporaryValue = array[currentIndex]
-        newArray[currentIndex] = array[randomIndex]
-        newArray[randomIndex] = temporaryValue
-    }
-
-    return newArray
 }
